Create a models folder
